refactor: bar tab 컴포넌트 tab으로 네이밍 수정 (#56)
