+++ src/shared/components/tab/tab/styles/tab-style.ts
