+++ src/shared/components/tab/tab/tab.stories.tsx
@@ -1,4 +1,4 @@
-import BarTabList from '@components/tab/bar-tab/bar-tab-list';
+import BarTabList from '@components/tab/tab/tab-list';
 import type { Meta, StoryObj } from '@storybook/react-vite';
 
 const meta: Meta<typeof BarTabList> = {
